refactor: move downloadLinkPlugin back to app/lib
